chore: extract components
